// src/generated/server/megabrain-market/conflict/v1/service_server.ts
export interface GeoCoordinates {
  latitude: number;
  longitude: number;
}

export interface PaginationResponse {
  nextCursor: string;
  totalCount: number;
}

export interface ListUcdpEventsRequest {
  start: number;
  end: number;
  pageSize: number;
  cursor: string;
  country: string;
}

export interface ListUcdpEventsResponse {
  events: UcdpViolenceEvent[];
  pagination?: PaginationResponse;
}

export interface UcdpViolenceEvent {
  id: string;
  dateStart: number;
  dateEnd: number;
  location?: GeoCoordinates;
  country: string;
  sideA: string;
  sideB: string;
  deathsBest: number;
  deathsLow: number;
  deathsHigh: number;
  violenceType: UcdpViolenceType;
  sourceOriginal: string;
}
// ...
export type UcdpViolenceType = "UCDP_VIOLENCE_TYPE_UNSPECIFIED" | "UCDP_VIOLENCE_TYPE_STATE_BASED" | "UCDP_VIOLENCE_TYPE_NON_STATE" | "UCDP_VIOLENCE_TYPE_ONE_SIDED";
// ...
export interface ServerContext {
  request: Request;
  pathParams: Record<string, string>;
  headers: Record<string, string>;
}